css/selectors/focus-visible-005.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8" />
  <title>CSS Test (Selectors): Programmatic focus causes :focus-visible to match</title>
  <link rel="author" title="Alice Boxhall" href="aboxhall@chromium.org" />
  <link rel="help" href="https://drafts.csswg.org/selectors-4/#the-focus-visible-pseudo" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <style>
    :root {
      --focus-light: hsl(50, 94%, 72%);
      --focus-visible-light: hsl(208, 56%, 78%);
      --focus-visible-dark: hsl(208, 56%, 70%);
    }

    :focus-visible {
      background: repeating-linear-gradient(
        45deg,
        var(--focus-visible-light),
        var(--focus-visible-light) 10px,
        var(--focus-visible-dark) 10px,
        var(--focus-visible-dark) 20px
      );
      background-color: var(--focus-visible-light);
    }

    :focus:not(:focus-visible) {
      background-color: var(--focus-light);
    }
  </style>
</head>
<body>
  This test checks that programmatically focusing an element causes <code>:focus-visible</code> matching to trigger.
  <ol id="instructions">
    <li>If the user-agent does not claim to support the <code>:focus-visible</code> pseudo-class then SKIP this test.</li>
    <li>Click the button below that says "Click me."</li>
    <li>If the element that says "I will be focused programmatically." does not have a striped blue background, then the test result is FAILURE. If the element <em>has</em> a striped blue background, then the test result is SUCCESS.</li>
  </ol>
  <br />
  <button id="button">Click me.</button>
  <div id="el" tabindex="-1">I will be focused programmatically.</div>
  <script>
    button.addEventListener("click", () => {
        el.focus();
    });
    async_test(function(t) {
        el.addEventListener("focus", t.step_func(() => {
            assert_equals(getComputedStyle(el).backgroundColor,
                          "rgb(167, 201, 230)");
            t.done();
        }));
        el.focus();
    }, "Programmatic focus should always match :focus-visible");
  </script>
</body>
</html>
